+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to the sdc11073 module will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.20] - 2022-12-12
+### Fixed
+- AllowedValue.Value is not optional
+- AlertSystemState LastSelfCheck and SelfCheckCount are only updated in self check cycle, 
+  not when part of transaction due to an AlertConditionState change
+
 ## [1.1.19] - 2022-12-08
 ### Fixed
 - iteration over DPWSHosted.types
-
 
 ## [1.1.18] - 2022-11-29
 ### Added